push code

+++ app/src/main/java/com/fussyvegan/scanner/model/ProductReview.java
@@ -95,7 +95,7 @@ public class ProductReview   implements Parcelable {
         this.review = review;
     }
 
-    public int getRating() {
+    public float getRating() {
         return rating;
     }
